+++ packages/api/src/functions/get.ts
@@ -8,7 +8,7 @@ import {
   getReferences,
   buildLookupPipeline,
   fill,
-} from '../collection'
+} from '../collection/index.js'
 
 export type GetOptions = {
   bypassSecurity?: boolean