+++ packages/access-control/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sonata-api/access-control
 
+## 1.0.20
+
+### Patch Changes
+
+- Fix esm builds
+- Updated dependencies
+  - @sonata-api/entrypoint@0.0.17
+  - @sonata-api/common@0.1.24
+  - @sonata-api/types@0.1.45
+
+## 1.0.19
+
+### Patch Changes
+
+- Entrypoint import fix
+- Updated dependencies
+  - @sonata-api/entrypoint@0.0.16
+  - @sonata-api/common@0.1.23
+  - @sonata-api/types@0.1.44
+
 ## 1.0.18
 
 ### Patch Changes