+++ packages/builtins/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @sonata-api/system
 
+## 1.0.31
+
+### Patch Changes
+
+- Fix esm builds
+- Updated dependencies
+  - @sonata-api/access-control@1.0.20
+  - @sonata-api/entrypoint@0.0.17
+  - @sonata-api/validation@1.0.19
+  - @sonata-api/common@0.1.24
+  - @sonata-api/types@0.1.45
+  - @sonata-api/api@1.0.45
+
+## 1.0.30
+
+### Patch Changes
+
+- Entrypoint import fix
+- Updated dependencies
+  - @sonata-api/entrypoint@0.0.16
+  - @sonata-api/access-control@1.0.19
+  - @sonata-api/api@1.0.44
+  - @sonata-api/common@0.1.23
+  - @sonata-api/types@0.1.44
+  - @sonata-api/validation@1.0.18
+
 ## 1.0.29
 
 ### Patch Changes