+++ packages/security/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sonata-api/security
 
+## 1.0.19
+
+### Patch Changes
+
+- Fix esm builds
+- Updated dependencies
+  - @sonata-api/common@0.1.24
+  - @sonata-api/types@0.1.45
+  - @sonata-api/api@1.0.45
+
+## 1.0.18
+
+### Patch Changes
+
+- Entrypoint import fix
+- Updated dependencies
+  - @sonata-api/api@1.0.44
+  - @sonata-api/common@0.1.23
+  - @sonata-api/types@0.1.44
+
 ## 1.0.17
 
 ### Patch Changes