+++ packages/server/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @sonata-api/server
 
+## 1.0.23
+
+### Patch Changes
+
+- Fix esm builds
+- Updated dependencies
+  - @sonata-api/access-control@1.0.20
+  - @sonata-api/entrypoint@0.0.17
+  - @sonata-api/builtins@1.0.31
+  - @sonata-api/common@0.1.24
+  - @sonata-api/server@1.0.23
+  - @sonata-api/types@0.1.45
+  - @sonata-api/http@1.0.24
+  - @sonata-api/api@1.0.45
+
+## 1.0.22
+
+### Patch Changes
+
+- Entrypoint import fix
+- Updated dependencies
+  - @sonata-api/entrypoint@0.0.16
+  - @sonata-api/access-control@1.0.19
+  - @sonata-api/api@1.0.44
+  - @sonata-api/builtins@1.0.30
+  - @sonata-api/common@0.1.23
+  - @sonata-api/http@1.0.23
+  - @sonata-api/server@1.0.22
+  - @sonata-api/types@0.1.44
+
 ## 1.0.21
 
 ### Patch Changes