+++ packages/sonata-build/CHANGELOG.md
@@ -1,5 +1,27 @@
 # sonata-build
 
+## 1.0.21
+
+### Patch Changes
+
+- Fix esm builds
+- Updated dependencies
+  - @sonata-api/builtins@1.0.31
+  - @sonata-api/common@0.1.24
+  - @sonata-api/types@0.1.45
+  - @sonata-api/api@1.0.45
+
+## 1.0.20
+
+### Patch Changes
+
+- Entrypoint import fix
+- Updated dependencies
+  - @sonata-api/api@1.0.44
+  - @sonata-api/builtins@1.0.30
+  - @sonata-api/common@0.1.23
+  - @sonata-api/types@0.1.44
+
 ## 1.0.19
 
 ### Patch Changes