+++ packages/types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sonata-api/types
 
+## 0.1.45
+
+### Patch Changes
+
+- Fix esm builds
+
+## 0.1.44
+
+### Patch Changes
+
+- Entrypoint import fix
+
 ## 0.1.43
 
 ### Patch Changes